\name{JPEN-package}
\alias{JPEN-package}
\alias{JPEN}
\docType{package}
\title{
\packageTitle{JPEN}
}
\description{\packageDescription{JPEN}}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{JPEN}
\packageIndices{JPEN}
}
\author{Ashwini Maurya, Email: mauryaas@msu.edu.
\packageAuthor{JPEN}
Maintainer: \packageMaintainer{JPEN}
}
\references{
A Well Conditioned and Sparse Estimate of Covariance and Inverse Covariance Matrix Using Joint Penalty. Submitted.
http://arxiv.org/pdf/1412.7907v2.pdf
}
\seealso{
jpen,jpen.inv}
